[#32] Update Restructring
